requirements.txt
# external requirements
click
Sphinx
coverage
#awscli>=1.29.0
flake8
pandas
#logging
numpy 
pathlib 
scikit-learn
imbalanced-learn
joblib
bentoml
#sys
pydantic
requests
pyjwt
pytest
